// crm_frontend_3/src/api/api.ts
export interface Order_Client_DetailsVm {
    iD_Order: number;
    name_Client: string;
    lastName_Client: string;
    email_Client?: string | undefined;
    telefon: string;
    type_technology: string;
    model_technology: string;
    breaking_info: string;
    quipment_info: string ;
    status_Order?: string | undefined;
    iD_Personnel_dispatcher: string;
    iD_Personnel_master: string;
    receipt_date: Date;
    issue_date?: Date | undefined;
    price?: number;
}
// ...
export interface Update_Order_ClientDto {
    iD_Order:number;
    status_Order: string ;
}